refactor: update `Axis` trait to include some methods so that the `nearest_one` methods can be identical between `float` and `fixed`. Refactor all of the nearest_one methods so that they use a common macro.
feat: implement `kiddo::fixed::kdtree::ArchivedKdTree::nearest_one
